Fully updated hero at all breakpoints
