+++ support-frontend/assets/helpers/theGrid.js
@@ -111,6 +111,7 @@ export const imageCatalogue: {
   printCampaigniMovo: 'f257e33b922f09ca180cc9205e0c766c2c2d529c/28_0_694_694',
   printCampaignHD: '817936f0d1a2755c778b523b5ac5daa2d2f27449/10_0_716_694',
   printCheckoutiMovo: '15677cf8f82a5690bec39bf73b2c098c8f79b24f/15_0_667_400',
+  editionsPackshot: '77c3a3756748137080863ef8e6dd690740f5a96b/0_0_1160_945',
 };
 
 // Utility type: https://flow.org/en/docs/types/utilities/#toc-keys
